test: Move dump_analysis() to dump.cpp
